pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "certsf"
version = "0.1.0"
description = "Certified special-function wrappers with explicit diagnostics."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mpmath>=1.3",
    "scipy>=1.10",
]

[project.optional-dependencies]
certified = [
    "python-flint>=0.6",
]
mcp = [
    "mcp>=1.0",
]
test = [
    "pytest>=8",
]
dev = [
    "mcp>=1.0",
    "pytest>=8",
    "python-flint>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"
